requirements.txt
Django==4.2.0
fastapi==0.95.2
uvicorn==0.23.0
pydantic==1.10.8
sqlalchemy==2.0.0
python-multipart==0.0.6
pillow==10.0.0
django-bootstrap5==23.3
djangorestframework==3.14.0
pandas==2.0.0
reportlab==3.6.12
django-tables2==2.5.0
whitenoise==6.5.0
gunicorn==21.2.0
celery==5.3.0
redis==4.6.0
django-crispy-forms==2.0.0
crispy-bootstrap5==0.7
xhtml2pdf==0.2.11
django-filter==23.2
django-widget-tweaks==1.4.12
django-cors-headers==4.0.0
python-dateutil==2.8.2
openpyxl==3.1.2
django-import-export==3.2.0
markdown==3.4.3
django-colorfield==0.9.0
django-admin-rangefilter==0.10.0
django-storages==1.13.2
WeasyPrint==59.0
python-slugify==8.0.1
six==1.16.0
psycopg2-binary==2.9.6 